+++ tosho_kmkc/src/constants.rs
@@ -31,6 +31,8 @@ pub struct Constants {
     pub(crate) hash: &'static str,
     /// Platform number
     pub(crate) platform_number: u8,
+    /// Support secure?
+    pub(crate) support_secure: bool,
 }
 
 /// A ranking tab for KM.
@@ -56,9 +58,10 @@ pub static ANDROID_CONSTANTS: LazyLock<Constants> = LazyLock::new(|| Constants {
     image_ua: "okhttp/4.9.3",
     platform: "2",
     version: "6.3.5",
-    display_version: Some("2.3.5"),
+    display_version: Some("3.0.0"),
     hash: HASH_HEADER_MOBILE,
     platform_number: 2,
+    support_secure: true,
 });
 /// The constants used for legacy Android devices.
 ///
@@ -71,6 +74,7 @@ pub static ANDROID_LEGACY_CONSTANTS: LazyLock<Constants> = LazyLock::new(|| Cons
     display_version: Some("2.1.5"),
     hash: HASH_HEADER_MOBILE,
     platform_number: 2,
+    support_secure: false,
 });
 /// The constants used for iOS devices.
 pub static APPLE_CONSTANTS: LazyLock<Constants> = LazyLock::new(|| {
@@ -89,6 +93,7 @@ pub static APPLE_CONSTANTS: LazyLock<Constants> = LazyLock::new(|| {
         display_version: None,
         hash: hash_header,
         platform_number: 1,
+        support_secure: false,
     }
 });
 /// The constants used for web devices.
@@ -104,11 +109,14 @@ pub static WEB_CONSTANTS: LazyLock<Constants> = LazyLock::new(|| {
         display_version: None,
         hash: hash_header,
         platform_number: 3,
+        support_secure: false,
     }
 });
 
 /// The base API used for overall requests.
 pub const BASE_API: &str = comptime_b64!("aHR0cHM6Ly9hcGkua21hbmdhLmtvZGFuc2hhLmNvbQ==");
+/// The secure base API used for overall requests.
+pub const SECURE_BASE_API: &str = comptime_b64!("aHR0cHM6Ly9zZS1hcGkua21hbmdhLmtvZGFuc2hhLmNvbQ==");
 /// The base image URL used for image requests.
 pub const BASE_IMG: &str = comptime_b64!("aHR0cHM6Ly9jZG4ua21hbmdhLmtvZGFuc2hhLmNvbQ==");